requirements.txt
flask
flask-cors
python-dotenv
google-genai
pypdf
chromadb
langchain
langchain-community
langchain-text-splitters
sentence-transformers
